Conflicts:
	src/Battlescape/TileEngine.cpp
	src/Ruleset/Ruleset.cpp
	src/Ruleset/Ruleset.h

+++ src/Engine/SurfaceSet.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2013 OpenXcom Developers.
+ * Copyright 2010-2014 OpenXcom Developers.
  *
  * This file is part of OpenXcom.
  *
@@ -60,7 +60,7 @@ public:
 	/// Gets the height of all frames.
 	int getHeight() const;
 	/// Gets the total frames in the set.
-	int getTotalFrames() const;
+	size_t getTotalFrames() const;
 	/// Sets the surface set's palette.
 	void setPalette(SDL_Color *colors, int firstcolor = 0, int ncolors = 256);
 	std::map<int, Surface*> *getFrames();
